refactor(events): extract shared event-table column builders
Pull the Type, Date, and Place column definitions out of EventsPage into
$lib/event-columns so the person Events tab can reuse them, keeping the
row-link accessibility pattern (stop-propagation, plain-text link) in one
place. Date sorting is opt-in so the tree-wide table keeps its chronological
order.
